Add native http benchmark
